traininSync-back/requirements.txt
asgiref==3.8.1
astroid==3.3.9
cfgv==3.4.0
coverage==7.8.2
dill==0.4.0
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
gunicorn==23.0.0
identify==2.6.10
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
uritemplate==4.1.1
virtualenv==20.31.1
whitenoise==6.9.0
